requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
numpy>=1.26.0
streamlit>=1.28.0